requirements.txt
annotated-types   0.7.0
anyio             4.9.0
click             8.2.1
dnspython         2.7.0
email_validator   2.2.0
exceptiongroup    1.3.0
fastapi           0.115.12
h11               0.16.0
idna              3.10
motor             3.7.1
pip               22.0.2
pydantic          2.11.5
pydantic_core     2.33.2
pymongo           4.13.0
python-dotenv     1.1.0
setuptools        59.6.0
sniffio           1.3.1
starlette         0.46.2
typing_extensions 4.14.0
typing-inspection 0.4.1
uvicorn           0.34.3